fixed #641
Former-commit-id: 127445ee9df78dc54c0a39ae550effa58c6185e3
Former-commit-id: 800b385a0a155c71ed89c42f51f59c06241c3ca3

+++ src/engine/ui.cpp
@@ -191,7 +191,7 @@ struct gui : guient
     //use to set min size (useful when you have progress bars)
     void strut(float size) { layout(isvertical() ? int(size*guibound[0]) : 0, isvertical() ? 0 : int(size*guibound[1])); }
     //add space between list items
-    void space(float size) { layout(isvertical() ? 0 : size*guibound[0], isvertical() ? size*guibound[1] : 0); }
+    void space(float size) { layout(isvertical() ? 0 : int(size*guibound[0]), isvertical() ? int(size*guibound[1]) : 0); }
 
     void pushfont(const char *font) { ::pushfont(font); fontdepth++; }
     void popfont() { if(fontdepth) { ::popfont(); fontdepth--; } }
